pyproject.toml
version = "0.1.0"
readme = "README.md"

[tool.aerich]
tortoise_orm = "app.api.db.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[project]
name = "fastapiminiproject"
version = "0.1.0"
dependencies = [
    "asgi-lifespan>=2.1.0",
    "email-validator>=2.3.0",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "aerich==0.7.1",
    "asgi-lifespan>=2.1.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
